refactor: move WindowFunctionInvocation into Expression
